refactor?(api): move read models to its bounded context
