+++ api/tests/tooling/domain-builder/factory/build-campaign-participation-info.js
@@ -1,4 +1,4 @@
-import { CampaignParticipationInfo } from '../../../../src/shared/domain/read-models/CampaignParticipationInfo.js';
+import { CampaignParticipationInfo } from '../../../../src/prescription/campaign/domain/read-models/CampaignParticipationInfo.js';
 
 function buildCampaignParticipationInfo({
   participantFirstName = 'participantFirstName',
